+++ spec/lightstreamer/errors_spec.rb
@@ -17,10 +17,19 @@ end
 
 describe Lightstreamer::RequestError do
   it 'constructs from an error message and error code' do
-    error = Lightstreamer::RequestError.new 'message', '404'
+    error = Lightstreamer::RequestError.new 'message', 404
 
     expect(error.request_error_message).to eq('message')
     expect(error.request_error_code).to eq(404)
+    expect(error.message).to eq('404: message')
+  end
+
+  it 'constructs from an error message' do
+    error = Lightstreamer::RequestError.new 'message', 0
+
+    expect(error.request_error_message).to eq('message')
+    expect(error.request_error_code).to eq(0)
+    expect(error.message).to eq('message')
   end
 end
 
@@ -63,7 +72,7 @@ describe Lightstreamer::Error do
   end
 
   it 'builds a base error when the numeric code is unknown' do
-    expect(Lightstreamer::Error).to receive(:new).with('message')
+    expect(Lightstreamer::Error).to receive(:new).with('999: message')
 
     Lightstreamer::Error.build 'message', '999'
   end