+++ spec/lightstreamer/subscription_spec.rb
@@ -34,6 +34,7 @@ describe Lightstreamer::Subscription do
 
     expect(subscription.process_stream_data("#{subscription.id},1|a|b")).to be true
     expect(subscription.process_stream_data("#{subscription.id},2|c|")).to be true
+    expect(subscription.process_stream_data("#{subscription.id},2,OV2")).to be true
     expect(subscription.process_stream_data('0,3|d|e')).to be false
 
     expect(calls.count).to eq(4)
@@ -74,13 +75,5 @@ describe Lightstreamer::Subscription do
 
     expect(subscription.retrieve_item_data(:item1)).to eq({})
     expect(subscription.retrieve_item_data(:item2)).to eq({})
-  end
-
-  it 'reports exceptions that occur in data callbacks' do
-    subscription.add_data_callback { raise 'test' }
-
-    expect do
-      subscription.process_stream_data "#{subscription.id},1|a|b"
-    end.to output("Lightstreamer: exception occurred in a subscription data callback: test\n").to_stderr
   end
 end